refactor(cli): act on the review of the workload renderer
Wrap the tab writer flush, so a failed table write says which write failed, the
way renderDefinitions already does.

Move the specs to the internal package, per the guideline that a test file sits
in the package it covers. The two generator test files #309 merged use the
external package and still need the same change.
